auth-service/requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
bcrypt>=4.0.0
alembic>=1.11.0
pytest>=7.3.0
httpx>=0.24.0 